pyproject.toml
[tool.poetry]
name = "mental-health-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Dominik Juelg <dominik.juelg@petleo.net>"]
readme = "README.md"
packages = [{include = "mental_health_telegram_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
